% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/condensed_index_from_row_col.R
\name{condensed_index_from_row_col}
\alias{condensed_index_from_row_col}
\title{Convert a row/col index to a condensed index (for use with dist).}
\usage{
condensed_index_from_row_col(i, j, n)
}
\arguments{
\item{i}{The row index of the uncondensed square matrix s.t. i!=j.}

\item{j}{The col index of the uncondensed square matrix s.t. i!=j.}

\item{n}{The length/width of the (square) matrix (i.e. the number of
observations passed into dist).}
}
\value{
A integer value giving the corresponding index into the condensed
  matrix (vector).
}
\description{
Convert a row/col index to a condensed index (for use with dist).
}
\examples{
data(iris)
n = nrow(iris)
dx = dist(iris[,-5])
i = 14; j=119

ix = condensed_index_from_row_col(i,j,n) # 1964
dist(iris[c(i,j),-5]) == dx[ix] # TRUE
}
\seealso{
\code{\link{dist}}, \code{\link{row_col_from_condensed_index}}
}

