% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/row_col_from_condensed_index.R
\name{row_col_from_condensed_index}
\alias{row_col_from_condensed_index}
\title{Convert a condensed index to a row/col index (for use with dist).}
\usage{
row_col_from_condensed_index(n, ix)
}
\arguments{
\item{n}{The length/width of the (square) matrix (i.e. the number of
observations passed into dist).}

\item{ix}{The condensed index to be converted}
}
\value{
a row and column index into the uncondensed square matrix.
}
\description{
Convert a condensed index to a row/col index (for use with dist).
}
\examples{
data(iris)
n = nrow(iris)
dx = dist(iris[,-5])
ix = which.max(dx) # 1964

rc_ix = row_col_from_condensed_index(n, ix) # 14, 119
dist(iris[rc_ix,-5]) == max(dx) # TRUE
}
\seealso{
\code{\link{dist}}, \code{\link{condensed_index_from_row_col}}
}

